Add solution for problem `A - Packets` Link: https://codeforces.com/contest/1037/problem/A Tags: constructive algorithms, greedy, math, *1300 Ref: https://codeforces.com/contest/1037/submission/42684332

+++ codeforces/1037/A.cpp
@@ -101,7 +101,7 @@ const ld PI=acos(-1.0);
 
 
 //memory consumption test
-ll a[(int)1e7];
+int a[(int)1e7];
 int main()
 {
     fast;
